Compare Tuples based on the element, not the element & score
Set uses equals to compare if the set is already in the set. Redis
Sorted Sets are unique on the value, not the value with the score. The
Tuple.equals should return true when the elements are equal regardless
of what the score is. The score should be used when comparing to another
element in the set to see if it is < or >.