Throw exception when trying to use jedis instance in Pipeline
To call same jedis instance which is in pipline makes no sense as
commands already sent to the pipeline are waiting to be processes. This
will cause inconsistencies whenever trying to execute commands.

A JedisDataException will be raised if client tries to use an instance
which is in Pipeline state. Aditionally `resetState` method can be
called to reset jedis instance and use it again normally.