Disable JMX of GenericObjectPoolConfig
When running 32 threads and every thread has a JedisCluster(), the
program will spend more than 30 seconds to init 32 JedisCluster()
instances.
The reason is in commons-pool-2.4 :

    GenericObjectPool() --> BaseGenericObjectPool() --> jmxRegister()

the jmxRegister() will spend a lot of time to register MBean because
it always use the same Instance name and have to loop many turns.

Because users can't affort minutes to init multi-instances of
JedisCluster(), I suggest to defaultly disable JMX in JedisCluster().