Problem
The current behavior of ShardedJedisPool is destroying the ShardedJedis instance once it is detected invalid. It impacts the performance when a shard is broken, because connections to the healthy shards will also be closed, and the client will have to communicate with the healthy shards with short lived connections.

Solution

Instead of abandon the ShardedJedis instance, we can just replace the broken connections.