format codes & extract an abstract cluster cache layer
The JedisClusterInfoCache has been reverted to the original one.
Add a new AbstractJedisCluserInfoCache.
If the user choose to use ReadFrom.MASTER, the old
JedisClusterInfoCache will be created.
If the user choose ReadFrom.SLAVE or BOTH, the
EnhancedJedisClusterInfoCache will be created.
EnhancedJedisClusterInfoCache can remove JedisPools in the process of
nodes setup