Fix binary value in comment in test code (#3402)
* Fix binary value in comment in test code src/test/java/redis/clients/jedis/commands/jedis/BitCommandsTest.java

When first setting "0" (string) to redis, the bits are actually 0011 0000, instead of number 0 with bits 0000 0000.

Goods new is: all the code logic are correct.   The problem only exists in the comment.

For convenience, I have some helper code to print bits and their positions (10-based), which makes bit navigation easier.
e.g. The final bits are:

```
0011 0001  0000 0100  0000 0000  0000 0000  0000 0001  
0123 4567  8901 2345  6789 0123  4567 8901  2345 6789  
0          1          2          3          4          
```