Do not rely on Redis compiled from source code
Issue #3710

Instead of building Redis server from source code, use Docker images.
Prepare docker-compose stacks that accept the version of the Docker
image as parameter, so we can run against multiple versions in CI.

The purpose of this change is to strictly move everything to Docker,
without trying to change anything else. Some Java tests had to be
adapted, still, because now the Redis instances are accessed via
localhost, while internally in the docker-compose network they see each
other with other IPs and ports.

Polish a bit the Makefile, for example to run exactly the same thing
that runs in CI, so when we run the tests locally we get exactly the
same coverage.