Streamline unit test execution in CI vs. locally
The current way of running JUnit tests has some drawbacks.

The CI executes tests by running multiple `mvn` commands. When running
tests locally, it is tedious to run the same multiple commands, in
order, for example, to get the same coverage as the CI gets. Ideally
the tests are configured in the project itself, in the pom.xml
file and by annotations on the tests, so that simply running `mvn test`
will do everything needed.

In order to achieve this, adapt the tests to be parameterized, so that
JUnit takes care of running them with different RESP protocols. It is OK
to use environment variables in general, for example for injecting
connection details. But it is not OK to use environment variables for
driving test execution, by injecting a RESP version and running the
tests multiple times. If I have to run more than `mvn test`, it does not
feel right.

Some tests rely on static fields and @BeforeClass and @AfterClass
annotations. These do not work well with parameterized tests, so remove
the static and switch to @Before and @After annotations.

The Makefile is a bit unpolished, i.e. the sequence of tasks is
controlled manually when it could be controlled declaratively, and the
`start` and `stop` tasks do not cover starting and stopping the Redis
Stack docker. Take care of these aspects.

All in all, now if I run `make test` I know I get the correct coverage
in the JaCoCo report. And the CI can also just run `make test`, without
additional work.