Reenable clustering tests (#3764)
* Reenable clustering tests

Uncomment clustering tests and adapt the code in the minimum possible
way to make these tests run.

* Include RESP protocol in cluster tests

The cluster tests were not using the configured RESP protocol, so adjust
that.

* Use a separate cluster for the bulk of clustering tests

Most of the clustering tests don't need to change the structure of the
cluster. Use for these tests a stable cluster, that is configured
outside the Java code and keeps its structure. The tests only flush the
data on the cluster nodes, which is safe and fast.

* Don't stunnel the stable cluster nodes

Co-authored-by: M Sazzadul Hoque <7600764+sazzad16@users.noreply.github.com>

* Move port range for stable cluster

* Move some tests into subclasses

The goal is to leave unchanged the top level test classes.

---------

Co-authored-by: Gabriel Erzse <gabriel.erzse@redis.com>
Co-authored-by: M Sazzadul Hoque <7600764+sazzad16@users.noreply.github.com>