Extensive unit tests for PipeliningBase (#3778)
* Extensive unit tests for PipeliningBase

Add extensive unit tests for the PipeliningBase abstract class. Since
this class holds a lot of functionality for its subclasses, it makes
sense to test it in isolation. And since the functionality is very
repetitive, it makes sense to use mocking. Basically this abstract class
just shuffles commands around from a CommandObjects to its subclasses,
and this can be covered very well with mocked tests.

* Fix flaky test

A slow log test seems to sometimes get a duration of zero for an
operation. Cover this case too in the assertion.

---------

Co-authored-by: Gabriel Erzse <gabriel.erzse@redis.com>