Add extensive tests for UnifiedJedis
Add mocked tests for UnifiedJedis. Most of the code is covered, except
the constructors and some iterator methods. Those are left for later.

Take the opportunity to split PipeliningTests into smaller test classes.
The idea is to have one superclass (MockCommandObjectsTest) that exposes
a variety of mocked CommandObjects. From there we branch in two
directions: towards PipeliningBase and towards UnifiedJedis. Each has a
test superclass that prepares mocked instances, and a set of concrete
test classes that contain the actual tests.

The tests are grouped in categories. The names of the categories are the
same as used on the redis.io website in the commands documentation page.
Inside each section (i.e. test class), the tested commands are ordered
alphabetically. The tests run in pairs, one for the string-based
command, and one for the equivalent byte-array-based command. Hopefully
this gives a good structure for following the code.

A new public constructor is needed in UnifiedJedis, in order to inject
the mocks.

Also, due to method visibility, one test class for UnifiedJedis must
reside in the same package as the class itself.