Extensive unit tests for the CommandObjects class
Write isolated unit test for the different methods of the CommandObjects
class. These can be tested by themselves, without considerations like
cluster, sentinel or other orthogonal aspects.

Structure the tests into logical groups, similar to how the Redis
commands are documented on the redis.io website. For each group of
commands there is a corresponding test class.

The main goal is to have the test code as readable as possible. The
methods are kept to a minimum length in general. Sometimes, due to
required data preparation, this is not possible.

Another goal is to have this as a sort of documentation. Ideally someone
who reads these tests should be able to grasp what the Redis commands
are doing.

This being said, the tests still focus on testing the client. For
example negative cases are rarely included, and mostly when they impact
the client, for example the parsing of the response. It is not the goal
of these tests to test the Redis server itself.