Add extensive tests for UnifiedJedis (#3788)
* Align Pipelined commands with UnifiedJedis

Add some Pipelined commands that are available in UnifiedJedis but are
missing the pipelined version.

* Add extensive tests for UnifiedJedis

Add mocked tests for UnifiedJedis. Most of the code is covered, except
the constructors and some iterator methods. Those are left for later.

Take the opportunity to split PipeliningTests into smaller test classes.
The idea is to have one superclass (MockCommandObjectsTest) that exposes
a variety of mocked CommandObjects. From there we branch in two
directions: towards PipeliningBase and towards UnifiedJedis. Each has a
test superclass that prepares mocked instances, and a set of concrete
test classes that contain the actual tests.

The tests are grouped in categories. The names of the categories are the
same as used on the redis.io website in the commands documentation page.
Inside each section (i.e. test class), the tested commands are ordered
alphabetically. The tests run in pairs, one for the string-based
command, and one for the equivalent byte-array-based command. Hopefully
this gives a good structure for following the code.

A new public constructor is needed in UnifiedJedis, in order to inject
the mocks.

Also, due to method visibility, one test class for UnifiedJedis must
reside in the same package as the class itself.

* Rename to subject under test to jedis

* Undo script commands changes

* Remove tests for removed code

* Mark new constructor as visible for testing

* React to code review

Rename classes as suggested. Move some tests into better suited classed.
Remove everything Graph related.

* Fix unit tests

And add two tests for recently added code.

---------

Co-authored-by: Gabriel Erzse <gabriel.erzse@redis.com>
Co-authored-by: M Sazzadul Hoque <7600764+sazzad16@users.noreply.github.com>