Make JSON imports optional in OSGi environments
When deploying in OSGi containers the Jedis bundle has two hard requirements on having
the GSON and JSON bundles around. Since these depenendencies are optional we mark
the imports as such in the OSGi manifest.

If these packages are not exported the bundle will be installed and will be unable to
make use of the JSON functionality. When they are present they will be used.

Additionally, allow newer version of the org.json bundle to be used without requiring
a new release of Jedis.